Seperate file for ark_serialize::test
